Separate generic and specific aspects of warning diagnostics
